Increase the update_xl sleep time to 20 secondsm
Earlier sleep time of 2 seconds was not sufficient for the full URL to load thus the extracted coordinates were imprecise. Increasing the sleep time to 20 seconds gurantees that the exact exact coordinates extracted. 20 seconds is likely an overkill but the script only needs to be run once , so the additional delay is acceptable.